rename: user domain 관련 파일들 경로 수정
