+++ src/main/java/yteamserver/domain/bookmark/application/BookmarkService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import yteamserver.domain.bookmark.domain.VideoBookmark;
 import yteamserver.domain.bookmark.domain.repository.VideoBookmarkRepository;
-import yteamserver.domain.users.repository.UserRepository;
-import yteamserver.domain.users.repository.Users;
+import yteamserver.domain.users.domain.repository.UserRepository;
+import yteamserver.domain.users.domain.Users;
 import yteamserver.domain.video.domain.Video;
 import yteamserver.domain.video.domain.repository.VideoRepository;
 import yteamserver.global.error.ErrorCode;
