+++ src/main/java/yteamserver/domain/users/domain/repository/UserCharacterRepository.java
@@ -1,6 +1,8 @@
-package yteamserver.domain.users.repository;
+package yteamserver.domain.users.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import yteamserver.domain.users.domain.Users;
+import yteamserver.domain.users.domain.UsersCharacters;
 
 import java.util.Optional;
