+++ src/main/java/yteamserver/domain/video/application/VideoService.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import yteamserver.domain.store.domain.Store;
 import yteamserver.domain.store.domain.repository.StoreRepository;
-import yteamserver.domain.users.repository.UserRepository;
-import yteamserver.domain.users.repository.Users;
+import yteamserver.domain.users.domain.repository.UserRepository;
+import yteamserver.domain.users.domain.Users;
 import yteamserver.domain.video.domain.Video;
 import yteamserver.domain.video.domain.repository.VideoRepository;
 import yteamserver.domain.video.dto.CreateVideoReq;
